Separate src/elona/util.hpp into src/util/{strutil,fileutil}.hpp
