+++ src/elona/autopick.cpp
@@ -1,8 +1,9 @@
 #include "autopick.hpp"
+#include "../util/fileutil.hpp"
+#include "../util/strutil.hpp"
 #include "data/types/type_item.hpp"
 #include "elona.hpp"
 #include "i18n.hpp"
-#include "util.hpp"
 #include "variables.hpp"
