+++ src/elona/character.cpp
@@ -1,6 +1,8 @@
 #include "character.hpp"
 #include <cassert>
 #include <type_traits>
+#include "../util/fileutil.hpp"
+#include "../util/strutil.hpp"
 #include "ability.hpp"
 #include "area.hpp"
 #include "buff.hpp"
