+++ src/elona/ctrl_inventory.cpp
@@ -1,3 +1,4 @@
+#include "../util/strutil.hpp"
 #include "ability.hpp"
 #include "audio.hpp"
 #include "calc.hpp"
