+++ src/elona/keybind/keybind.cpp
@@ -1,4 +1,5 @@
 #include "keybind.hpp"
+#include "../../util/strutil.hpp"
 #include "../enums.hpp"
 #include "../gdata.hpp"
 #include "../log.hpp"
