+++ src/elona/lua_env/lua_console.cpp
@@ -3,12 +3,12 @@
 #include "../../snail/application.hpp"
 #include "../../snail/blend_mode.hpp"
 #include "../../snail/input.hpp"
+#include "../../util/strutil.hpp"
 #include "../config/config.hpp"
 #include "../filesystem.hpp"
 #include "../input.hpp"
 #include "../macro.hpp"
 #include "../ui.hpp"
-#include "../util.hpp"
 #include "../variables.hpp"
 #include "api_manager.hpp"
 #include "lua_env.hpp"
