+++ src/elona/ui.cpp
@@ -1,4 +1,5 @@
 #include "ui.hpp"
+#include "../util/strutil.hpp"
 #include "ability.hpp"
 #include "audio.hpp"
 #include "character.hpp"
