+++ src/elona/ui/ui_menu_feats.cpp
@@ -1,5 +1,6 @@
 #include "ui_menu_feats.hpp"
 #include "../../snail/color.hpp"
+#include "../../util/strutil.hpp"
 #include "../enchantment.hpp"
 #include "../menu.hpp"
 #include "../message.hpp"
